requirements.txt
asgiref==3.8.1
attrs==24.2.0
cfgv==3.4.0
distlib==0.3.8
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
Faker==28.0.0
filelock==3.15.4
identify==2.6.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pre-commit==3.8.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
virtualenv==20.26.3
